PGPRO-2032: Fix log-rotation-size and log-rotation-age integer overflow

+++ src/pg_probackup.h
@@ -191,8 +191,8 @@ typedef struct pgBackupConfig
 	char	   *log_filename;
 	char	   *error_log_filename;
 	char	   *log_directory;
-	int			log_rotation_size;
-	int			log_rotation_age;
+	uint64		log_rotation_size;
+	uint64		log_rotation_age;
 
 	uint32		retention_redundancy;
 	uint32		retention_window;
